introduce trip simulation read method
